Small fixes and adding ascii art

+++ src/utils/extra_configs.py
@@ -33,5 +33,5 @@ class SubtitleBackground(BaseModel):
     radius: int
     offset: int
     transformy: int
-    height_scaling: int
+    height_scaling: float
     rounded_border: bool
